+++ wkpnbot/__init__.py
@@ -1,4 +1,4 @@
 __author__ = "wkpn"
 __license__ = "MIT"
 __url__ = "https://github.com/wkpn/wkpnbot"
-__version__ = "2.2.0"
+__version__ = "2.3.0"