+++ wkpnbot/routers/forum/__init__.py
@@ -1,4 +1,8 @@
 from .actions import build_forum_actions_router
+from .commands import build_forum_commands_router
 
 
-__all__ = ["build_forum_actions_router"]
+__all__ = [
+    "build_forum_actions_router",
+    "build_forum_commands_router"
+]