+++ wkpnbot/routers/forum/helpers.py
@@ -1,0 +1,6 @@
+from aiogram.filters.command import Command
+
+
+class CommandWipeForumTopic(Command):
+    def __init__(self):
+        super().__init__("wipe")