+++ wkpnbot/utils/__init__.py
@@ -1,20 +1,26 @@
 from .callbacks import (
     EditedMessageCallback,
-    UserUpdateCallback
+    UserUpdateCallback,
+    WipeForumTopicCallback
 )
+from .defaults import TRASH_EMOJI_ID
 from .keyboards import (
     build_edited_message_keyboard,
     build_privacy_policy_keyboard,
     build_user_card_keyboard,
+    build_wipe_forum_topic_keyboard
 )
 from .user_card import  make_user_card_info
 
 
 __all__ = [
+    "TRASH_EMOJI_ID",
     "EditedMessageCallback",
     "UserUpdateCallback",
+    "WipeForumTopicCallback",
     "build_edited_message_keyboard",
     "build_privacy_policy_keyboard",
     "build_user_card_keyboard",
+    "build_wipe_forum_topic_keyboard",
     "make_user_card_info"
 ]