+++ wkpnbot/utils/defaults.py
@@ -1,2 +1,3 @@
 DEFAULT_PHOTO = "https://i.imgur.com/fAXIQ5K.png"
-DEFAULT_PRIVACY_POLICY_LINK = "https://telegra.ph/Privacy-Policy-07-03-60"
+PRIVACY_POLICY_LINK = "https://telegra.ph/Privacy-Policy-07-03-60"
+TRASH_EMOJI_ID = "5260337865237222404"