+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 2.3.0 - August 16th, 2024
+
+### Added
+
+- `/wipe` command to delete current forum topic and messages data from the database.
+- `DBClient.delete`, `DBClient.delete_many`, `DBClient.fetch_many` methods for database client.
+- `CallbackQuery` support for `TopicsManagementMiddleware`.
+- Additional `content_type` checks for `FilterMiddleware`.
+- Additional database check before constructing `reply_parameters` in `MessagesMiddleware`.
+- Support for deleting messages with a special premium reaction (forum only).
+
+### Changed
+
+- Update typings and imports.
+- Update `aiogram` to 3.12.0.
+- Update `tomlkit` to 0.13.2.
+- Update `DBClient` attribute names.
+
 ## 2.2.0 - August 9th, 2024
 
 ### Changed
@@ -33,7 +51,7 @@
 
 ### Added
 
-- Full support of [aiogram](https://github.com/aiogram) v3.
+- Full support of [aiogram](https://github.com/aiogram/aiogram) v3.
 - `pyproject.toml` support.
 - Webhook configuration script (`configure_webhook.py`).
 - Webhook handler (`wkpnbot/webserver.py`)